Q&N Chat bot/requirements.txt
streamlit
pillow
pdf2image
google-generativeai
python-dotenv
